backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.11
django-colorfield==0.11.0
django-filter==24.2
django-import-export==3.3.8
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.6
import-export==0.3.1
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
